fix(llm): remove unnecessary cast in temperature parameter
- Remove redundant double cast from temperature parameter in OllamaChatOptions
- Temperature value is already a double type, casting is unnecessary
- Improves code clarity and reduces type conversion overhead

+++ cortex-llm/src/main/java/be/jsilkens/cortex/llm/SpringAiLlmAdapter.java
@@ -21,7 +21,7 @@ public class SpringAiLlmAdapter implements LlmPort {
         try {
             var options = OllamaChatOptions.builder()
                     .model(request.getModel())
-                    .temperature((double) request.getTemperature())
+                    .temperature(request.getTemperature())
                     .maxTokens(request.getMaxTokens())
                     .build();
